// src/generated/openapi.d.ts
export interface paths {
    "/tinkoff.public.invest.api.contract.v1.InstrumentsService/BondBy": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** Метод получения облигации по её идентификатору. */
        post: operations["InstrumentsService_BondBy"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /**
         * @default RUB
         * @enum {string}
         */
        PortfolioRequestCurrencyRequest: "RUB" | "USD" | "EUR";
// ...
        rpcStatus: {
            /** Format: int32 */
            code?: number;
            message?: string;
            details?: components["schemas"]["protobufAny"][];
        };
// ...
        v1CandleInterval: "CANDLE_INTERVAL_UNSPECIFIED" | "CANDLE_INTERVAL_1_MIN" | "CANDLE_INTERVAL_5_MIN" | "CANDLE_INTERVAL_15_MIN" | "CANDLE_INTERVAL_HOUR" | "CANDLE_INTERVAL_DAY" | "CANDLE_INTERVAL_2_MIN" | "CANDLE_INTERVAL_3_MIN" | "CANDLE_INTERVAL_10_MIN" | "CANDLE_INTERVAL_30_MIN" | "CANDLE_INTERVAL_2_HOUR" | "CANDLE_INTERVAL_4_HOUR" | "CANDLE_INTERVAL_WEEK" | "CANDLE_INTERVAL_MONTH";
// ...
        v1InstrumentIdType: "INSTRUMENT_ID_UNSPECIFIED" | "INSTRUMENT_ID_TYPE_FIGI" | "INSTRUMENT_ID_TYPE_TICKER" | "INSTRUMENT_ID_TYPE_UID" | "INSTRUMENT_ID_TYPE_POSITION_UID";
// ...
        v1InstrumentType: "INSTRUMENT_TYPE_UNSPECIFIED" | "INSTRUMENT_TYPE_BOND" | "INSTRUMENT_TYPE_SHARE" | "INSTRUMENT_TYPE_CURRENCY" | "INSTRUMENT_TYPE_ETF" | "INSTRUMENT_TYPE_FUTURES" | "INSTRUMENT_TYPE_SP" | "INSTRUMENT_TYPE_OPTION" | "INSTRUMENT_TYPE_CLEARING_CERTIFICATE";
// ...
        v1OperationState: "OPERATION_STATE_UNSPECIFIED" | "OPERATION_STATE_EXECUTED" | "OPERATION_STATE_CANCELED" | "OPERATION_STATE_PROGRESS";
// ...
        v1OrderDirection: "ORDER_DIRECTION_UNSPECIFIED" | "ORDER_DIRECTION_BUY" | "ORDER_DIRECTION_SELL";
// ...
        v1OrderType: "ORDER_TYPE_UNSPECIFIED" | "ORDER_TYPE_LIMIT" | "ORDER_TYPE_MARKET" | "ORDER_TYPE_BESTPRICE";
// ...
        v1Quotation: {
            /**
             * целая часть суммы, может быть отрицательным числом
             * Format: int64
             */
            units?: string;
            /**
             * дробная часть суммы, может быть отрицательным числом
             * Format: int32
             */
            nano?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}